FEAT: implements commitments for InternalNode
Start adding support for commitments in the Trie.
InternalNode commits to its children's commitments.
On inserting a new element in the Trie, the affected Nodes become dirty.
On setCommitment, it recomputes the commitment if isDirty.

Also, setup commitments for LeafNodes, but remains to implement.
Setup update_commitment in the IpaMultipoint JNI, also needs implementation.