ADD: adds HashVisitor + refactors Nodes and Visitors a little.
HashVisitor is a new feature not present for MPT.
We choose to follow the visitor pattern for all algorithms including computing hashes.
Previously were implemented get and put, next will be remove and commit.

Changes to Nodes: we allow inserting many values before updating hashes.
The HashVisitor is responsible to update all hashes and can be called when desired.
Ultimately, hash will need to be present before use, but it can be optional empty to mark that it should be computed.
No need for an extra flag like dirty for persistance.

Also, we are moving an extra step towards immutability of nodes: hash is final, constructors were updated.

Visitors follow a uniform interface: take a node, possibly extra arguments, and returns a node.
Previously, order of arguments was not consistent, and it could return void.